Three tools needed to run the project: FLTK, glew and glem. Way to set them:
While the Base Project has everything you need to start an OpenGL/FLTK project, sometimes (ok…most of the time) your setup at home will be different than the labs. In the labs, FLTK and GLM were located in Program Files (x86) while Glew was in Program Files for some reason. It’s important to know how to set project properties so Visual Studio recognizes the libraries.

There are four relevant changes to make to the Project Properties (right click the project name -> properties):

Make sure the “Configuration” setting at the top left is set to “All Configurations” and not “Active(Debug),” so your changes apply to both the Debug and Release project configurations.
In the C++(General) menu, edit the “Additional Include Directories” to point to where you have the FLTK, GLM, and Glew include files. See the base project for examples.
In the Linker(General) menu, edit the “Additional Library Dependencies” in a similar way to point to the relevant .lib files.
In the Linker(Input) menu, add the names of the needed library files to the “Additional Dependencies” field. Usually, these will be some FLTK and Glew libraries, but it will vary with each project.